[Change] Serializers separation.
