+++ virtualenv/SaaS/api/serializers/skill.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from ..models.skill import Skill
+
+
+class SkillSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Skill
+        fields = ('id', 'name')
